split tests out for markers. wrote utility to render
